+++ include/vpd.h
@@ -25,14 +25,14 @@ struct machine_info {
 const struct machine_info *machine_info_lookup(const char *mtm);
 
 const void *vpd_find_keyword(const void *rec, size_t rec_sz,
-			     const char *kw, uint8_t *kw_size);
+			     const char *kw, uint16_t *kw_size);
 
 const void *vpd_find_record(const void *vpd, size_t vpd_size,
 			    const char *record, size_t *sz);
 
 const void *vpd_find(const void *vpd, size_t vpd_size,
 		     const char *record, const char *keyword,
-		     uint8_t *sz);
+		     uint16_t *sz);
 
 bool vpd_valid(const void *vvpd, size_t vpd_size);
